ML-Backend/requirements.txt
fastapi
youtube-transcript-api
langchain
langchain-community
langchain-google-genai
google-generativeai
chromadb
tiktoken 
python-dotenv
faiss-cpu
